appendix/requirements.txt
Django==1.7.2
django-appconf==0.6
django-compressor==1.4
django-filter==0.9.2
djangorestframework==3.0.3
python-dateutil==2.4.0
six==1.9.0
